Reverse and head glitch
Ok first fixed some of the reverse nonsense. Then fixed the head glitch issue. The head follow the mouse which cause it to detached oddly great for testing but unneeded now. Added a extra 0 point for the tail so angle constraints are easier and added two for the head the first for angle constraint and the second for the mouse so the fish move towards it.